requirements.txt
django==2.2
django-braces==1.14.0
pillow==8.3.1
djangorestframework==3.12.4